+++ functions/.eslintrc.json
@@ -76,30 +76,20 @@
       {
         "skipWordIfMatch": [
           "aqi",
-          "timestamps",
           "firebase",
           "Firebase",
           "firestore",
           "Firestore",
-          "thingspeak",
-          "Thingspeak",
           "axios",
           "Axios",
-          "csv",
-          "Csv",
           "pubsub",
           "Aqi",
           "nowcast",
-          "fs",
-          "os",
-          "tmpdir",
-          "Atm",
           "camelcase",
           "subcollection",
           "Subcollection",
           "breakpoint",
           "Breakpoint",
-          "uid",
           "purpleair",
           "Recurse",
           "readonly"
