+++ functions/src/get-reading/util.ts
@@ -1,5 +1,3 @@
-import {PurpleAirReading} from './types';
-
 /**
  * Converts PurpleAir's confidence value into percent difference
  * @param confidence - confidence value from PurpleAir, between 0 and 100
@@ -29,109 +27,22 @@ import {PurpleAirReading} from './types';
 function getMeanPercentDifference(confidence: number): number {
   const minConfidence = 0;
   const maxConfidence = 100;
-  /* eslint-disable no-magic-numbers */
+
+  const minPercentDifference = 0;
+  const maxPercentDifference = 2;
   switch (confidence) {
     case minConfidence:
       // If the confidence is zero, then we return the maximum possible percent difference
-      return 2;
+      return maxPercentDifference;
     case maxConfidence:
       // The confidence value from PurpleAir can be 100 even if channel A and
       // channel B do not completely match, but if the confidence value is 100,
       // then the value is data is good enough to meet the EPA recommendation.
-      return 0;
+      return minPercentDifference;
     default:
       // Otherwise, undo the calculation from the PurpleAir confidence value
-      /* eslint-disable-next-line no-magic-numbers */
+      // eslint-disable-next-line no-magic-numbers
       return ((maxConfidence - confidence + 25) * 1.6) / 100;
-  }
-}
-
-/**
- * Converts a PurpleAir reading returned from the group API query into a PurpleAirReading
- * @param data - list of data from PurpleAir for a given sensor
- * @param fieldNames - list of field names from PurpleAir that match the order of the data fields
- * @returns
- */
-function getReading(
-  data: (string | number)[],
-  fieldNames: string[]
-): [number, PurpleAirReading | null] {
-  // Initialize all values
-  let id: number = Number.NaN;
-  let name: string | undefined = undefined;
-  let latitude: number | undefined = undefined;
-  let longitude: number | undefined = undefined;
-  let meanPercentDifference: number | undefined = undefined;
-  let pm25: number | undefined = undefined;
-  let humidity: number | undefined = undefined;
-  let timestamp: Date | undefined = undefined;
-
-  data.forEach((value, index) => {
-    // Check the corresponding field name to determine how to handle the value
-    switch (fieldNames[index]) {
-      case 'sensor_index':
-        if (typeof value === 'number') id = value;
-        break;
-      case 'name':
-        if (typeof value === 'string') name = value;
-        break;
-      case 'latitude': {
-        if (typeof value === 'number') latitude = value;
-        break;
-      }
-      case 'longitude':
-        if (typeof value === 'number') longitude = value;
-        break;
-      case 'confidence':
-        if (typeof value === 'number') {
-          meanPercentDifference = getMeanPercentDifference(value);
-        }
-        break;
-      case 'pm2.5':
-        if (typeof value === 'number') pm25 = value;
-        break;
-      case 'humidity':
-        if (typeof value === 'number') humidity = value;
-        break;
-      case 'last_seen':
-        if (typeof value === 'number') {
-          // PurpleAir returns seconds since EPOCH, but the Date constructor
-          // takes milliseconds, so we convert from seconds to milliseconds
-          timestamp = new Date(value * 1000); // eslint-disable-line no-magic-numbers
-        }
-        break;
-      default:
-        // Unknown field, ignore
-        break;
-    }
-  });
-
-  // Only return a PurpleAirReading if all fields are defined
-  if (
-    id &&
-    name &&
-    latitude !== undefined && // Can be zero
-    longitude !== undefined && // Can be zero
-    meanPercentDifference !== undefined && // Can be zero
-    pm25 !== undefined && // Can be zero
-    humidity !== undefined && // Can be zero
-    timestamp
-  ) {
-    return [
-      id,
-      {
-        id: id,
-        name: name,
-        latitude: latitude,
-        longitude: longitude,
-        meanPercentDifference: meanPercentDifference,
-        pm25: pm25,
-        humidity: humidity,
-        timestamp: timestamp,
-      },
-    ];
-  } else {
-    return [id, null];
   }
 }
 
@@ -158,4 +69,4 @@ async function getLastSensorReadingTime(
   return lastSensorReadingTime;
 }
 
-export {getLastSensorReadingTime, getReading, getMeanPercentDifference};
+export {getLastSensorReadingTime, getMeanPercentDifference};
