+++ src/components/Admin/ManageSensors/DownloadData/Util.tsx
@@ -85,7 +85,10 @@ const MonthInput: ({value, setValue}: InputProps) => JSX.Element = ({
     t('downloadData.months.nov'),
     t('downloadData.months.dec'),
   ];
-  const options = [];
+  const options: React.DetailedHTMLProps<
+    React.OptionHTMLAttributes<HTMLOptionElement>,
+    HTMLOptionElement
+  >[] = [];
   for (let i = 1; i < labels.length; i++) {
     options.push(
       <option value={i} key={i}>
@@ -119,7 +122,10 @@ const DayInput: ({value, setValue}: InputProps) => JSX.Element = ({
   setValue,
 }: InputProps) => {
   const {t} = useTranslation('administration');
-  const options = [];
+  const options: React.DetailedHTMLProps<
+    React.OptionHTMLAttributes<HTMLOptionElement>,
+    HTMLOptionElement
+  >[] = [];
   const maxDaysPerMonth = 31;
   for (let i = 0; i < maxDaysPerMonth; i++) {
     options.push(
