+++ src/components/Util/Dates.ts
@@ -36,7 +36,8 @@ function formatTime(hour: number, minutes: number): string {
     hour = hour % hoursPerPeriod;
   }
   let leadingZero = '';
-  /* eslint-disable-next-line no-magic-numbers */
+
+  // eslint-disable-next-line no-magic-numbers
   if (minutes < 10) {
     leadingZero = '0';
   }
